Fix minor issue in UTF-32 decoder.
In streaming mode, if the first chunk is < 32 bytes, but there are more
chunks written with total size of the stream > 32 bytes, then we were
losing initial chunk. This looks like an unlikely scenario, so not a
major issue.

Note, this is happening only in UTF-32 decoder, not UTF-32LE or BE,
as the problem was in the code that detects encoding.